# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Cross Mule Backend API
  version: 1.0.0
  description: REST API for the Cross Mule Detection dashboard frontend.
servers:
  - url: http://localhost:8000
    description: Local server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        refreshToken:
          type: string
        user:
          type: object
          properties:
            id:
              type: string
            email:
              type: string
            name:
              type: string
            organization:
              type: string
    Account:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        riskScore:
          type: number
        type:
          type: string
          enum: [suspicious, normal, high-risk]
        channels:
          type: array
          items:
            type: string
        balance:
          type: number
        clusterId:
          type: string
          nullable: true
        jurisdiction:
          type: string
        owner:
          type: string
        confidenceScore:
          type: number
          nullable: true
        sanctionsMatch:
          type: boolean
    Transaction:
      type: object
      properties:
        id:
          type: string
        fromAccount:
          type: string
        toAccount:
          type: string
        amount:
          type: number
        channel:
          type: string
          enum: [UPI, ATM, Wallet, App, Web]
        timestamp:
          type: string
          format: date-time
        riskScore:
          type: number
        status:
          type: string
          enum: [completed, pending, flagged]
        pattern:
          type: string
          nullable: true
        complexity:
          type: number
          nullable: true
        sanctionsFlag:
          type: boolean
security:
  - bearerAuth: []
paths:
  /health:
    get:
      summary: Service health check
      security: []
      responses:
        '200':
          description: Service status

  /api/auth/signup:
    post:
      summary: Create user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
                name:
                  type: string
                fullName:
                  type: string
                organization:
                  type: string
      responses:
        '201':
          description: Signup success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  /api/auth/login:
    post:
      summary: Login with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  /api/auth/refresh:
    post:
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                refreshToken:
                  type: string
      responses:
        '200':
          description: New access token

  /api/auth/logout:
    post:
      summary: Logout and revoke refresh token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                refreshToken:
                  type: string
      responses:
        '200':
          description: Logout success

  /api/accounts:
    get:
      summary: List accounts
      responses:
        '200':
          description: Accounts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
    post:
      summary: Create account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Account'
      responses:
        '201':
          description: Account created

  /api/accounts/{id}:
    get:
      summary: Get account by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Account
    put:
      summary: Update account by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated account
    delete:
      summary: Delete account by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deletion result

  /api/transactions:
    get:
      summary: List transactions
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: Transactions list
    post:
      summary: Create transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transaction'
      responses:
        '201':
          description: Transaction created

  /api/transactions/live:
    get:
      summary: Get latest live transactions
      responses:
        '200':
          description: Latest transactions

  /api/transactions/{id}:
    get:
      summary: Get transaction by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transaction
    put:
      summary: Update transaction by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated transaction
    delete:
      summary: Delete transaction by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deletion result

  /api/clusters:
    get:
      summary: List clusters
      responses:
        '200':
          description: Clusters list

  /api/clusters/{id}:
    get:
      summary: Get cluster by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cluster

  /api/alerts:
    get:
      summary: List alerts
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: severity
          schema:
            type: string
      responses:
        '200':
          description: Alerts list

  /api/alerts/{id}:
    get:
      summary: Get alert by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Alert

  /api/alerts/{id}/status:
    patch:
      summary: Update alert status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [new, investigating, resolved]
      responses:
        '200':
          description: Updated alert

  /api/alerts/{id}/dismiss:
    post:
      summary: Dismiss alert
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated alert

  /api/graph:
    get:
      summary: Graph nodes and links
      responses:
        '200':
          description: Graph data

  /api/graph/nodes:
    get:
      summary: Graph nodes
      responses:
        '200':
          description: Nodes list

  /api/graph/links:
    get:
      summary: Graph links
      responses:
        '200':
          description: Links list

  /api/risk/distribution:
    get:
      summary: Risk distribution buckets
      responses:
        '200':
          description: Distribution data

  /api/risk/jurisdictions:
    get:
      summary: Jurisdiction risk analytics
      responses:
        '200':
          description: Jurisdiction data

  /api/risk/ownership:
    get:
      summary: Ownership correlation analytics
      responses:
        '200':
          description: Ownership data

  /api/channels/flow:
    get:
      summary: Channel flow analytics
      responses:
        '200':
          description: Channel flow

  /api/channels/velocity:
    get:
      summary: Velocity trend analytics
      responses:
        '200':
          description: Velocity trend

  /api/reports/risk-trend:
    get:
      summary: Risk trend report data
      parameters:
        - in: query
          name: period
          schema:
            type: string
      responses:
        '200':
          description: Risk trend

  /api/reports/complexity:
    get:
      summary: Complexity report data
      responses:
        '200':
          description: Complexity data

  /api/reports/generate:
    post:
      summary: Generate report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reportType:
                  type: string
      responses:
        '201':
          description: Generated report id

  /api/reports/export:
    get:
      summary: Export report as csv or pdf
      parameters:
        - in: query
          name: reportId
          schema:
            type: string
        - in: query
          name: format
          schema:
            type: string
            enum: [csv, pdf]
      responses:
        '200':
          description: Exported file

  /api/intelligence/sharing:
    get:
      summary: List intelligence sharing partnerships
      responses:
        '200':
          description: Partnership list

  /api/intelligence/partnerships:
    post:
      summary: Create intelligence sharing partnership
      responses:
        '201':
          description: Created partnership

  /api/stats/dashboard:
    get:
      summary: Dashboard metrics
      responses:
        '200':
          description: Stats object

  /api/data:
    get:
      summary: Generic data endpoint
      responses:
        '200':
          description: Generic stats

  /api/submit:
    post:
      summary: Generic submit endpoint
      responses:
        '201':
          description: Submit acknowledgement